requirements.txt
langchain
langchain-community
bs4
arxiv
wikipedia
langchain-text-splitters
python-dotenv
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
chromadb
streamlit
langchain_groq
langchain_core  
fastapi
uvicorn
validators==0.28.1
youtube_transcript_api
unstructured
pytube
sentence-transformers
HuggingFaceHub